5258: responsive balance  page
