+++ src/ducks/HistoricalBalance/AssetsField.js
@@ -1,17 +1,8 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import gql from 'graphql-tag'
 import { graphql } from 'react-apollo'
 import { Select } from '@santiment-network/ui'
-
-const allAssetsGQL = gql`
-  {
-    allErc20Projects {
-      slug
-      mainContractAddress
-    }
-  }
-`
+import { allAssetsGQL } from './common/queries'
 
 class AssetsField extends Component {
   state = {
