+++ src/ducks/HistoricalBalance/chart/HistoricalBalanceChart.module.scss
@@ -1,0 +1,6 @@
+@import 'variables.scss';
+@import "~@santiment-network/ui/mixins.scss";
+
+.chartContainer {
+  height: 450px;
+}
