+++ src/ducks/HistoricalBalance/common/queries.js
@@ -20,3 +20,12 @@ export const historicalBalanceGQL = gql`
     }
   }
 `
+
+export const allAssetsGQL = gql`
+  {
+    allErc20Projects {
+      slug
+      mainContractAddress
+    }
+  }
+`
